requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
Pillow>=10.0.0
